Make it possible to excecute without LYMAN_DIR (primarily for writing graphs)

+++ lyman/frontend.py
@@ -1,6 +1,7 @@
 """Forward facing lyman tools with information about ecosystem."""
 import os
 import os.path as op
+import tempfile
 import re
 import sys
 import imp
@@ -191,7 +192,10 @@ def load_info_from_module(module_name, lyman_dir):
 def load_scan_info(lyman_dir=None):
     """Load information about subjects, sessions, and runs from scans.yaml."""
     if lyman_dir is None:
-        lyman_dir = os.environ["LYMAN_DIR"]
+        lyman_dir = os.environ.get("LYMAN_DIR", None)
+
+    if lyman_dir is None:
+        return {}
 
     scan_fname = op.join(lyman_dir, "scans.yaml")
     with open(scan_fname) as fid:
@@ -237,14 +241,15 @@ def info(experiment=None, model=None, lyman_dir=None):
 
     """
     if lyman_dir is None:
-        lyman_dir = os.environ["LYMAN_DIR"]
+        lyman_dir = os.environ.get("LYMAN_DIR", None)
 
     # --- Load project-level information
-    project_info = load_info_from_module("project", lyman_dir)
-    check_extra_vars(project_info, ProjectInfo)
-
-    # Load scan information
-    project_info["scan_info"] = load_scan_info(lyman_dir)
+    if lyman_dir is None:
+        project_info = {}
+    else:
+        project_info = load_info_from_module("project", lyman_dir)
+        check_extra_vars(project_info, ProjectInfo)
+        project_info["scan_info"] = load_scan_info(lyman_dir)
 
     # --- Load the experiment-level information
     if experiment is None:
@@ -273,9 +278,17 @@ def info(experiment=None, model=None, lyman_dir=None):
             .trait_set(**model_info))
 
     # Ensure that directories are specified as real absolute paths
+    if lyman_dir is None:
+        base = op.join(tempfile.mkdtemp(), "lyman")
+    else:
+        base = lyman_dir
     directories = ["data_dir", "proc_dir", "cache_dir"]
     orig = info.trait_get(directories)
-    full = {k: op.abspath(op.join(lyman_dir, v)) for k, v in orig.items()}
+    full = {k: op.abspath(op.join(base, v)) for k, v in orig.items()}
+    for d in full.values():
+        if not op.exists(d):
+            os.mkdir(d)
+
     info.trait_set(**full)
 
     return info
@@ -311,7 +324,10 @@ def subjects(subject_arg=None, sessions=None, lyman_dir=None):
     scan_info = load_scan_info(lyman_dir)
 
     if lyman_dir is None:
-        lyman_dir = os.environ["LYMAN_DIR"]
+        lyman_dir = os.environ.get("LYMAN_DIR", None)
+
+    if lyman_dir is None:
+        return []
 
     # -- Parse the input
 
@@ -413,7 +429,7 @@ def execute(wf, args, info):
     # After successful completion of the workflow, optionally delete the
     # intermediate files, which are not usually needed aside from debugging
     # (persistent outputs go into the `info.proc_dir`).
-    if info.remove_cache and not args.debug:
+    if info.remove_cache and not args.debug and op.exists(cache_dir):
         shutil.rmtree(cache_dir)
 
     return res
